requirements.txt
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
streamlit==1.21.0
scikit-learn==1.2.2
